Add role based functionality (#15) (#16)
* Setup file structure (#1)

* Modify .gitignore and run npm init

* Setup file structure

* Modify package.json and package-lock.json (#2)

* Modify package.json and package-lock.json

* Add configurations for TypeScript, Webpack, ESLint, and Prettier. Also add scripts to package.json (#3)

* Create server.ts, apiRoutes.ts, dataController.ts, and userController.ts (#5)

* Serve basic React frontend and configure backend to work with TS, Webpack, and ESLint (#6)

* Generate random combination and implement basic functionality for frontend components (#7)

* Create settings page and handle settings state through Context API (#8)

* Implement main game logic, guess history, and guess statistics (#9)

* Implement session management and session-store using express-session and connect-mongo (#10)

* Add signup and signin functionality (#11)

* Add user and admin roles

+++ src/server/controllers/userController.ts
@@ -8,7 +8,11 @@ const userController = {
       const { email, password } = req.body;
       const hashedPassword = await argon2.hash(password);
 
-      console.log('Creating user with data:', { email, password: hashedPassword });
+      // catch missing fields in the request body
+      if (!email || !password) {
+        return res.status(400).json({ message: 'Email and password are required' });
+      }
+
       await UserModel.create({ email, password: hashedPassword });
 
       return next();
@@ -20,6 +24,11 @@ const userController = {
   getUser: async (req: Request, res: Response, next: NextFunction) => {
     try {
       const { id } = req.params;
+      // catch missing id in the request params
+      if (!id) {
+        return res.status(400).json({ message: 'Missing id or malformed request' });
+      }
+
       const user = await UserModel.findById(id);
 
       res.locals.user = user;
@@ -45,6 +54,9 @@ const userController = {
   deleteUser: async (req: Request, res: Response, next: NextFunction) => {
     try {
       const { id } = req.params;
+      if (!id) {
+        return res.status(400).json({ message: 'Missing id or malformed request' });
+      }
 
       await UserModel.findByIdAndDelete(id);
